Update/Delete methods added to GraphQL endpoint

+++ src/Catalog.Api/Requests/ItemRequest.cs
@@ -12,7 +12,7 @@ namespace Catalog.Api.Requests
         [Required]
         public int CategoryId { get; set; }
         [Required]
-        public decimal Price { get; set; }
+        public float Price { get; set; }
         [Required]
         public int Amount { get; set; }
     }
